Move VideoPlayer to src/Components/VideoPlayer. Fix minor issues
