+++ src/Components/News/index.js
@@ -2,10 +2,11 @@ import Title from '../Titles';
 import NewsList from './NewsList';
 // import React, { useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
+import LastEvent from './LastEvent';
 // import { getAllNews } from '../../app/features/newsSlice';
 
 export const News = () => {
-  const { news } = useSelector(state => state);
+  const { news } = useSelector((state) => state);
   // const [newsList, setNewsList] = useState([]);
   // const [loading, setLoading] = useState(false);
   // const [error, setError] = useState(false);
@@ -29,9 +30,14 @@ export const News = () => {
   // }, [news]);
 
   return (
-        <>
-            <Title children="Novedades"/>
-            <NewsList newsList={news.news || []} loading={news.newsLoading} error={news.newsError}/>
-        </>
+    <>
+      <Title children='Novedades' />
+      <LastEvent video={'https://youtu.be/4YnSk1gI_Oo'} />
+      <NewsList
+        newsList={news.news || []}
+        loading={news.newsLoading}
+        error={news.newsError}
+      />
+    </>
   );
 };
